refactor: navigation-drawer themes
