--- src/components/nav-drawer/nav-drawer-header-item.indigo.scss
@@ -1,7 +1,0 @@
-@use '../../styles/utilities' as utils;
-
-@mixin theme() {
-    igc-nav-drawer-header-item {
-        color: #{utils.color(gray, 50)};
-    }
-}
