--- src/components/nav-drawer/nav-drawer-item.indigo.scss
@@ -1,26 +1,0 @@
-@use '../../styles/utilities' as utils;
-
-@mixin theme() {
-    igc-nav-drawer-item::part(base) {
-        color: #{utils.color(gray, 300, .8)};
-        border-radius: #{utils.rem(24px)};
-        padding: #{utils.rem(4px)};
-        margin: #{utils.rem(15px)} #{utils.rem(8px)} !important;
-
-        &:hover,
-        &:focus-within {
-            background: transparent;
-            color: #{utils.color(gray, 100)};
-        }
-    }
-
-    igc-nav-drawer-item::part(content) {
-        margin-inline-start: #{utils.rem(24px)};
-        padding-inline-start: #{utils.rem(8px)};
-    }
-
-    igc-nav-drawer-item[active]::part(base) {
-        color: #{utils.color(gray, 50)};
-        background: #{utils.color(gray, 600, .8)};
-    }
-}
