+++ src/components/nav-drawer/styles/fluent/nav-drawer-header-item.fluent.scss
@@ -1,0 +1,3 @@
+@use '../../../../styles/common/component';
+@use '../../../../styles/utilities/index' as *;
+@use '../nav-drawer-header-item.base';
