+++ src/components/nav-drawer/styles/indigo/nav-drawer-header-item.indigo.scss
@@ -1,0 +1,8 @@
+@use '../../../../styles/common/component';
+@use '../../../../styles/utilities/index' as *;
+@use '../nav-drawer-header-item.base';
+
+:host {
+    color: color(gray, 50);
+    background: red;
+}
