+++ src/components/nav-drawer/styles/indigo/nav-drawer-item.indigo.scss
@@ -1,0 +1,30 @@
+@use '../../../../styles/common/component';
+@use '../../../../styles/utilities/index' as *;
+@use '../nav-drawer-item.base';
+
+:host {
+    [part~='base'] {
+        color: color(gray, 300, .8);
+        border-radius: rem(24px);
+        padding: rem(4px);
+        margin: rem(15px) rem(8px) !important;
+
+        &:hover,
+        &:focus-within {
+            background: transparent;
+            color: color(gray, 100);
+        }
+    }
+
+    [part='content'] {
+        margin-inline-start: rem(24px);
+        padding-inline-start: rem(8px);
+    }
+}
+
+:host([active]) {
+    [part~='base'] {
+        background: color(gray, 600, .8);
+        color: color(gray, 50);
+    }
+}
