+++ src/components/nav-drawer/styles/nav-drawer-header-item.base.scss
@@ -1,5 +1,5 @@
-@use '../../styles/common/component';
-@use '../../styles/utilities' as *;
+@use '../../../styles/common/component';
+@use '../../../styles/utilities/index' as *;
 
 :host {
     display: block;
