+++ src/components/nav-drawer/styles/nav-drawer-item.base.scss
@@ -1,5 +1,5 @@
-@use '../../styles/common/component';
-@use '../../styles/utilities' as *;
+@use '../../../styles/common/component';
+@use '../../../styles/utilities/index' as *;
 
 $item-margin: rem(8px);
