+++ src/styles/themes/light/indigo.scss
@@ -7,9 +7,6 @@
 @use '../../../components/radio/radio.indigo' as radio;
 @use '../../../components/checkbox/switch.indigo' as switch;
 @use '../../../components/navbar/navbar.indigo' as navbar;
-@use '../../../components/nav-drawer/nav-drawer.indigo' as nav-drawer;
-@use '../../../components/nav-drawer/nav-drawer-item.indigo' as nav-drawer-item;
-@use '../../../components/nav-drawer/nav-drawer-header-item.indigo' as nav-drawer-header;
 @use '../../../components/list/list.indigo' as list;
 @use '../../../components/input/input.indigo' as input;
 @use '../../../components/slider/slider.indigo' as slider;
@@ -32,9 +29,6 @@ $palette: utils.palette(
 @include radio.theme();
 @include switch.theme();
 @include navbar.theme();
-@include nav-drawer.theme();
-@include nav-drawer-item.theme();
-@include nav-drawer-header.theme();
 @include list.theme();
 @include input.theme();
 @include icon-button.theme();
